refactor: move hooks from component dirs to src/hooks/
Move hooks that were incorrectly placed in component directories:
- useBinInspector.ts: inspector/ -> hooks/ (shared by RightPanel, MobileInspector)
- useBinGeometry.ts: IsometricPreview/ -> hooks/ (3D bin rendering)

Update all imports and re-export from original locations for backwards compat.
