+++ src/components/Grid/IsometricPreview/BinMesh.tsx
@@ -2,7 +2,7 @@ import { useRef } from 'react';
 import * as THREE from 'three';
 import { useFrame } from '@react-three/fiber';
 import type { Bin } from '../../../types';
-import { useBinGeometry } from './useBinGeometry';
+import { useBinGeometry } from '../../../hooks/useBinGeometry';
 
 interface BinMeshProps {
   bin: Bin;
