+++ src/hooks/useBinGeometry.ts
